Add AuthMerging directive to basic authentication
By default, authentication blocks are not merged with one-another with
the more specific block overriding the less specific one. In this
context this resulted in the .htaccess blocking in apache2 being
overriden by the basic authentication added by patchpanel making the
.htaccess files accessible on sites that have basic auth enabled.

Adding `AuthMerging And` results in the more specific block taking
the less-specific block into account.